feat: Add btc_received field to registration API response
The code changes in the registration API endpoint add a new field, `btc_received`, to the API response. This field is set to 0 for registrations with a status of "Pending" and is updated with the amount received from the BTC Pay Server for registrations with a status of "Complete". This enhancement provides additional information about the BTC received for each registration.