feat: Update event creation API route and form
This commit updates the API route for creating events to include a new parameter, "fee". The fee parameter is required in the request body and is used to set the fee for the event. Additionally, the "max_players" parameter has been fixed to always have a value of 100.

In the frontend, the event creation form has been updated to set the default date to one month from the current date and time. This change is implemented using the moment library.

This commit also includes some minor code cleanup and fixes in other files.